+++ ValheimServersideQoL/Processors/Processor.cs
@@ -511,7 +511,7 @@ abstract class Processor
         public static void ShowMessage(long targetPeerId, MessageHud.MessageType type, string message)
         {
             /// Invoke <see cref="MessageHud.RPC_ShowMessage"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(targetPeerId, "ShowMessage", (int)type, message);
+            ZRoutedRpc.instance.InvokeRoutedRPC(targetPeerId, "ShowMessage", parameters: [(int)type, message]);
         }
 
         //public static void ShowMessage(MessageHud.MessageType type, string message)
@@ -530,13 +530,13 @@ abstract class Processor
         {
             playerZdo.AssertIs<Player>();
             /// <see cref="Player.UseStamina(float)"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(playerZdo.GetOwner(), playerZdo.m_uid, "UseStamina", value);
+            ZRoutedRpc.instance.InvokeRoutedRPC(playerZdo.GetOwner(), playerZdo.m_uid, "UseStamina", parameters: [value]);
         }
 
         public static void SendGlobalKeys(Peer peer, List<string> keys)
         {
             /// <see cref="ZoneSystem.SendGlobalKeys"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(peer.m_uid, "GlobalKeys", keys);
+            ZRoutedRpc.instance.InvokeRoutedRPC(peer.m_uid, "GlobalKeys", parameters: [keys]);
         }
 
         public static void ShowInWorldText(IEnumerable<long> targetPeerIds, DamageText.TextType type, Vector3 pos, string text)
@@ -548,7 +548,7 @@ abstract class Processor
             zPackage.Write(text);
             zPackage.Write(false);
             foreach (var peer in targetPeerIds)
-                ZRoutedRpc.instance.InvokeRoutedRPC(peer, "RPC_DamageText", zPackage);
+                ZRoutedRpc.instance.InvokeRoutedRPC(peer, "RPC_DamageText", parameters: [zPackage]);
         }
 
         //public static void ShowInWorldText(IEnumerable<Peer> peers, DamageText.TextType type, Vector3 pos, string text)
@@ -563,7 +563,7 @@ abstract class Processor
         static void TeleportPlayer(long targetPeerID, Vector3 pos, Quaternion rot, bool distantTeleport)
         {
             /// <see cref="Chat.TeleportPlayer(long, Vector3, Quaternion, bool)"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(targetPeerID, "RPC_TeleportPlayer", pos, rot, distantTeleport);
+            ZRoutedRpc.instance.InvokeRoutedRPC(targetPeerID, "RPC_TeleportPlayer", parameters: [pos, rot, distantTeleport]);
         }
 
         public static void TeleportPlayer(Peer peer, Vector3 pos, Quaternion rot, bool distantTeleport)
@@ -573,21 +573,21 @@ abstract class Processor
         {
             player.AssertIs<Player>();
             /// <see cref="Player.TeleportTo(Vector3, Quaternion, bool)"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(player.GetOwner(), player.m_uid, "RPC_TeleportTo", pos, rot, distantTeleport);
+            ZRoutedRpc.instance.InvokeRoutedRPC(player.GetOwner(), player.m_uid, "RPC_TeleportTo", parameters: [pos, rot, distantTeleport]);
         }
 
         public static void Remove(ExtendedZDO piece, bool blockDrop = false)
         {
             piece.AssertIs<Piece>();
             /// <see cref="WearNTear.RPC_Remove"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(piece.GetOwner(), piece.m_uid, "RPC_Remove", false);
+            ZRoutedRpc.instance.InvokeRoutedRPC(piece.GetOwner(), piece.m_uid, "RPC_Remove", parameters: [false]);
         }
 
         public static void AddStatusEffect(ExtendedZDO character, int nameHash, bool resetTime = false, int itemLevel = 0, float skillLevel = 0f)
         {
             character.AssertIs<Character>();
             /// <see cref="SEMan.AddStatusEffect"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(character.GetOwner(), character.m_uid, "RPC_AddStatusEffect", [nameHash, resetTime, itemLevel, skillLevel]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(character.GetOwner(), character.m_uid, "RPC_AddStatusEffect", parameters: [nameHash, resetTime, itemLevel, skillLevel]);
         }
 
         public static void RequestStack(ExtendedZDO container, ExtendedZDO player, long playerID = 0)
@@ -598,7 +598,7 @@ abstract class Processor
             /// <see cref="Container.RPC_RequestStack"/>
             if (playerID is 0)
                 playerID = player.Vars.GetPlayerID();
-            InvokeRoutedRPCAsSender(player.GetOwner(), container.GetOwner(), container.m_uid, "RPC_RequestStack", [playerID]);
+            InvokeRoutedRPCAsSender(player.GetOwner(), container.GetOwner(), container.m_uid, "RPC_RequestStack", parameters: [playerID]);
         }
 
         public static void StackResponse(ExtendedZDO container, bool granted)
@@ -606,14 +606,14 @@ abstract class Processor
             container.AssertIs<Container>();
 
             /// <see cref="Container.RPC_StackResponse"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "RPC_StackResponse", [granted]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "RPC_StackResponse", parameters: [granted]);
         }
 
         public static void TakeAllResponse(ExtendedZDO container, bool granted)
         {
             container.AssertIs<Container>();
             /// <see cref="Container.RPC_TakeAllRespons"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "TakeAllRespons", [granted]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "TakeAllRespons", parameters: [granted]);
         }
 
         public static void RequestStateChange(ExtendedZDO trap, int state)
@@ -621,7 +621,7 @@ abstract class Processor
             trap.AssertIs<Trap>();
 
             /// <see cref="Trap.RPC_RequestStateChange"/>"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(trap.GetOwner(), trap.m_uid, "RPC_RequestStateChange", [state]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(trap.GetOwner(), trap.m_uid, "RPC_RequestStateChange", parameters: [state]);
         }
 
         public static void SetTamed(ExtendedZDO character, bool tamed)
@@ -629,7 +629,7 @@ abstract class Processor
             character.AssertIs<Character>();
 
             /// <see cref="Character.SetTamed(bool)"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(character.GetOwner(), character.m_uid, "RPC_SetTamed", [tamed]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(character.GetOwner(), character.m_uid, "RPC_SetTamed", parameters: [tamed]);
         }
 
         public static void Damage(ExtendedZDO character, HitData hitData)
@@ -637,7 +637,7 @@ abstract class Processor
             character.AssertIs<Character>();
 
             /// <see cref="Character.Damage(HitData)"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(character.GetOwner(), character.m_uid, "RPC_Damage", [hitData]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(character.GetOwner(), character.m_uid, "RPC_Damage", parameters: [hitData]);
         }
 
         public static void RequestOwn(ExtendedZDO itemDrop, [CallerFilePath] string callerFile = default!, [CallerLineNumber] int callerLineNo = default)
@@ -652,14 +652,21 @@ abstract class Processor
         {
             container.AssertIs<Container>();
             /// <see cref="Container.RPC_RequestOpen"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "RequestOpen", [playerID]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "RequestOpen", parameters: [playerID]);
         }
 
         public static void OpenResponse(ExtendedZDO container, bool granted)
         {
             container.AssertIs<Container>();
             /// <see cref="Container.RPC_OpenRespons"/>
-            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "OpenRespons", [granted]);
+            ZRoutedRpc.instance.InvokeRoutedRPC(container.GetOwner(), container.m_uid, "OpenRespons", parameters: [granted]);
+        }
+
+        public static void DamageMineRock5(ExtendedZDO minerock5, HitData hit, int hitAreaIndex)
+        {
+            minerock5.AssertIs<MineRock5>();
+            /// <see cref="MineRock5.RPC_Damage"/>
+            ZRoutedRpc.instance.InvokeRoutedRPC(minerock5.GetOwner(), minerock5.m_uid, "RPC_Damage", parameters: [hit, hitAreaIndex]);
         }
 
         static Action<ZRoutedRpc, long, ZDOID, string, object[], long>? __invokeRouteRPCAsSender;